+++ src/reductstore/index.ts
@@ -9,6 +9,11 @@ export { EXAMPLES } from './examples';
 export { DIRECTIVES } from './directives';
 export { getCompletionProvider } from './completionProvider';
 
-export { SQL_KEYWORDS, SQL_FUNCTIONS } from './sql/keywords';
-export { SQL_COMPARISON_OPERATORS } from './sql/operators';
+export { SQL_CLAUSES } from './sql/clauses';
+export { SQL_FUNCTIONS } from './sql/functions';
+export {
+  SQL_COMPARISON_OPERATORS,
+  SQL_LOGICAL_OPERATORS,
+  SQL_FILTER_OPERATORS,
+} from './sql/operators';
 export { getSqlCompletionProvider } from './sql/completionProvider';